+++ IRasRag.Domain/Entities/SpeciesStageConfig.cs
@@ -15,12 +15,12 @@ namespace IRasRag.Domain.Entities
         public Guid FeedTypeId { get; set; }
 
         [Required]
-        public float AmountPer100Fish { get; set; } // Recommended amount in kg
+        public double AmountPer100Fish { get; set; } // Recommended amount in kg
 
         [Required]
         public int FrequencyPerDay { get; set; } // How many times to feed per day
 
-        public float? MaxStockingDensity { get; set; } // Max fish per cubic meter (or liter) allowed in this stage
+        public double? MaxStockingDensity { get; set; } // Max fish per cubic meter (or liter) allowed in this stage
 
         public int? ExpectedDurationDays { get; set; } // How many days the fish typically stay in this stage
